+++ src/app/components/layout/auth/Background.tsx
